Separacion de archivos. Mucho mejor, a mi entender. Fixes #28
